refactor: move promisifyGetCSS to cssProperties.js
